Added dotenv package for database vars basic exeption handling

+++ main.py
@@ -1,11 +1,19 @@
-from fastapi import FastAPI
+import os
+from fastapi import FastAPI, HTTPException
 from sqlalchemy import create_engine
 from db_models import Person
 from api_models import Person as PersonRequestBody
 from sqlalchemy.orm import Session
+from dotenv import load_dotenv
+
+load_dotenv()
 
 app = FastAPI()
-engine = create_engine("postgresql://localhost/matt")
+
+try:
+    engine = create_engine(os.environ.get("DATABASE_URL"))
+except Exception as e:
+    print(e)
 
 
 @app.get("/people")
@@ -27,6 +35,8 @@ def add_person(person: PersonRequestBody):
 def delete_person(person_id: int):
     with Session(engine) as session:
         person = session.query(Person).filter(Person.id == person_id).first()
+        if person is None:
+            raise HTTPException(status_code=404, detail="Person not found")
         session.delete(person)
         session.commit()
         return "Deleted person"
